Reorganized Frechet into the R'n'D section because it is just that, R&D
