+++ hoot-core-test/hoot-core-test.pro
@@ -246,9 +246,7 @@ SOURCES += \
     src/test/cpp/hoot/core/algorithms/KskipBigramDistanceTest.cpp \
     src/test/cpp/hoot/core/test/ConflateCaseTest.cpp \
     src/test/cpp/hoot/core/test/ConflateCaseTestSuite.cpp \
-    src/test/cpp/hoot/core/algorithms/linearreference/WayStringTest.cpp \
-    src/test/cpp/hoot/core/algorithms/FrechetDistanceTest.cpp \
-    src/test/cpp/hoot/core/algorithms/FrechetSublineMatcherTest.cpp
+    src/test/cpp/hoot/core/algorithms/linearreference/WayStringTest.cpp 
 
 HEADERS += \
     src/test/cpp/hoot/core/test/ConflateCaseTest.h \
