+++ hoot-rnd/hoot-rnd.pro
@@ -59,7 +59,9 @@ SOURCES += \
     src/test/cpp/hoot/rnd/conflate/network/OsmNetworkTest.cpp \
     src/test/cpp/hoot/rnd/conflate/network/NetworkVertexTest.cpp \
     src/test/cpp/hoot/rnd/conflate/network/NetworkEdgeTest.cpp \
-    src/test/cpp/hoot/rnd/conflate/polygon/AreaMergerTest.cpp
+    src/test/cpp/hoot/rnd/conflate/polygon/AreaMergerTest.cpp \
+    src/test/cpp/hoot/rnd/conflate/frechet/FrechetDistanceTest.cpp \
+    src/test/cpp/hoot/rnd/conflate/frechet/FrechetSublineMatcherTest.cpp
 
 HEADERS += \
     src/test/cpp/hoot/rnd/conflate/network/LegacyVertexMatcherTest.h 
@@ -96,7 +98,9 @@ SOURCES += \
     src/main/cpp/hoot/rnd/conflate/network/EdgeSublineMatch.cpp \
     src/main/cpp/hoot/rnd/conflate/network/PartialNetworkMerger.cpp \
     src/main/cpp/hoot/rnd/conflate/network/ConflictsNetworkMatcher.cpp \
-    src/main/cpp/hoot/rnd/conflate/polygon/AreaMerger.cpp
+    src/main/cpp/hoot/rnd/conflate/polygon/AreaMerger.cpp \
+    src/main/cpp/hoot/rnd/conflate/frechet/FrechetDistance.cpp \
+    src/main/cpp/hoot/rnd/conflate/frechet/FrechetSublineMatcher.cpp 
 
 HEADERS += \
     src/main/cpp/hoot/rnd/ops/KeepPoiAndBuildingRefOp.h \
@@ -133,7 +137,9 @@ HEADERS += \
     src/main/cpp/hoot/rnd/conflate/network/PartialNetworkMerger.h \
     src/main/cpp/hoot/rnd/conflate/network/ConflictsNetworkMatcher.h \
     src/main/cpp/hoot/rnd/conflate/network/EidMapper.h \
-    src/main/cpp/hoot/rnd/conflate/polygon/AreaMerger.h
+    src/main/cpp/hoot/rnd/conflate/polygon/AreaMerger.h \
+    src/main/cpp/hoot/rnd/conflate/frechet/FrechetDistance.h \
+    src/main/cpp/hoot/rnd/conflate/frechet/FrechetSublineMatcher.h
 
 FORMS +=
