+++ hoot-rnd/src/main/cpp/hoot/rnd/conflate/frechet/FrechetDistance.cpp
