+++ hoot-rnd/src/test/cpp/hoot/rnd/conflate/frechet/FrechetDistanceTest.cpp
@@ -35,7 +35,7 @@
 #include <geos/geom/LineString.h>
 
 // Hoot
-#include <hoot/core/algorithms/FrechetDistance.h>
+#include <hoot/rnd/conflate/frechet/FrechetDistance.h>
 #include <hoot/core/visitors/FindWaysVisitor.h>
 
 #include <hoot/core/io/OsmReader.h>
@@ -43,7 +43,7 @@
 #include <hoot/core/io/OsmWriter.h>
 #include <hoot/core/MapProjector.h>
 
-#include "../TestUtils.h"
+#include <hoot-core-test/src/test/cpp/hoot/core/TestUtils.h>
 
 using namespace hoot;
